backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
flask-restful==0.3.9
requests
python-dotenv==1.0.1
nltk
numpy
scikit-learn
psycopg2-binary
docx2txt
